fixes warping bug with ray_plane intersection

+++ solution/inc/intersection.h
@@ -17,6 +17,6 @@
 
 int intersection_triangle_aabb(Triangle triangle, AABB bounding_box);
 int intersection_ray_aabb(Ray r, AABB box, double *tmin, double *tmax);
-int intersection_ray_plane(Ray r, Plane p);
+double intersection_ray_plane(Ray r, Plane p);
 
 #endif
